+++ subdaap/utils.py
@@ -4,6 +4,10 @@ import os
 
 
 class VerboseAction(argparse.Action):
+    """
+    Argparse action to count the verbose level (e.g. -v, -vv, etc).
+    """
+
     def __call__(self, parser, args, value, option_string=None):
         try:
             value = int(value or "1")
@@ -14,11 +18,19 @@ class VerboseAction(argparse.Action):
 
 
 class NewPathAction(argparse.Action):
+    """
+    Argparse action that resolves a given path to absolute path.
+    """
+
     def __call__(self, parser, args, values, option_string=None):
         setattr(args, self.dest, os.path.abspath(values))
 
 
 class PathAction(argparse.Action):
+    """
+    Argparse action that resolves a given path, and ensures it exists.
+    """
+
     def __call__(self, parser, args, values, option_string=None):
         path = os.path.abspath(values)